requirements.txt
# Core Django Framework
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<5.0
django-filter>=24.3,<25.0

# Database
psycopg2-binary>=2.9.10,<3.0
django-extensions>=3.2.3,<4.0

# Authentication & Security
djangorestframework-simplejwt>=5.4,<6.0
django-guardian>=2.4.0,<3.0

# API Documentation
drf-spectacular>=0.28,<0.29

# File Handling & Reports
Pillow>=11.0,<12.0
reportlab>=4.2,<5.0
WeasyPrint>=63.0,<64.0
openpyxl>=3.1.5,<4.0
xlsxwriter>=3.2,<4.0

# Barcode Generation
python-barcode>=0.15.1,<1.0
qrcode>=8.0,<9.0

# EDI Processing (optional - install separately if needed)
# pyx12==2.3.3  # Latest available version, uncomment if EDI processing is required

# Task Queue (for background jobs)
celery>=5.4,<6.0
redis>=5.2,<6.0
django-celery-beat>=2.7,<3.0

# Utilities
python-dateutil>=2.9,<3.0
requests>=2.32,<3.0

# Development
python-dotenv>=1.0.1,<2.0
ipython>=8.30,<9.0

# Testing
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
factory-boy>=3.3.1,<4.0

# Production
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
